+++ NEWS.md
@@ -1,3 +1,12 @@
+# lastdose 0.3.0
+
+- Add `time_col` and `id_col` arguments to let user name columns containing 
+  time and user id information #14
+- ID column is allowed to be character; data will be converted to numeric if 
+  it is character #14
+- Time column is allowed to be `POSIXct`; data will be converted to numeric 
+  according to value of `time_units` argument #9
+
 # lastdose 0.2.1
 - Fix bug where behavior is undefined when ADDL or II columns are omitted
   (#11)